requirements.txt
flask
pymysql
flask-sqlalchemy
flask_login
flask_bcrypt
pillow
pdfkit